+++ app/components/daisy/feedback/alert_component.rb
@@ -9,11 +9,21 @@
 # @part content_wrapper  [HTML] A wrapper for the main content of the alert.
 #   This allows for proper spacing and alignment with the icon.
 #
+# @part close            [Button] A close button that appears when closable is true.
+#
 # @option kws [String] :style (:default) The style of the alert.
 #   [:info, :success, :warning, :error, :default]
 # @option kws [Boolean] :soft (false) Use the soft style variant.
 # @option kws [Boolean] :outline (false) Use the outline style variant.
 # @option kws [Boolean] :dash (false) Use the dash style variant.
+# @option kws [Integer] :timeout (nil) Auto-dismiss timeout in milliseconds.
+#   If nil, uses global configuration default. If false, no auto-dismiss.
+# @option kws [Boolean] :autoclose (false) Enable auto-dismiss using timeout.
+#   Must be true for auto-dismiss to work.
+# @option kws [String] :href (nil) Converts alert to a clickable link.
+# @option kws [String] :action (nil) Stimulus action to fire on click.
+# @option kws [Boolean] :closable (false) Show close button. Set to true to enable
+#   manual dismissal.
 #
 # @loco_example Basic Alert
 #   = daisy_alert do
@@ -104,10 +114,31 @@
 #   = daisy_alert(css: "alert-error alert-dash") do
 #     This is a dash error alert.
 #
+# @loco_example Closable Alert
+#   = daisy_alert(icon: "information-circle", css: "alert-info", closable: true, icon_html: { variant: :outline }) do
+#     This alert can be closed manually.
+#
+# @loco_example Auto-dismissing Alert
+#   = daisy_alert(icon: "check-circle", css: "alert-success", autoclose: true, timeout: 3000, icon_html: { variant: :outline }) do
+#     This alert will auto-dismiss in 3 seconds.
+#
+# @loco_example Clickable Link Alert
+#   = daisy_alert(icon: "information-circle", css: "alert-info", href: "/docs", icon_html: { variant: :outline }) do
+#     Click to view documentation.
+#
+# @loco_example Stimulus Action Alert
+#   = daisy_alert(icon: "exclamation-triangle", css: "alert-warning", action: "click->my-controller#handle", icon_html: { variant: :outline }) do
+#     Click to trigger custom action.
+#
 class Daisy::Feedback::AlertComponent < LocoMotion::BaseComponent
   include LocoMotion::Concerns::IconableComponent
+  include LocoMotion::Concerns::LinkableComponent
 
-  define_parts :icon, :content_wrapper
+  define_parts :icon, :content_wrapper, :close
+
+  # @return [Boolean] Whether or not this alert can be closed.
+  attr_reader :closable
+  alias :closable? :closable
 
   #
   # Creates a new Alert component.
@@ -130,11 +161,23 @@ class Daisy::Feedback::AlertComponent < LocoMotion::BaseComponent
   # @option kws soft     [Boolean] Use the soft style variant.
   # @option kws outline  [Boolean] Use the outline style variant.
   # @option kws dash     [Boolean] Use the dash style variant.
+  # @option kws timeout  [Integer] Auto-dismiss timeout in milliseconds.
+  #   If nil, uses global configuration default. If false, no auto-dismiss.
+  # @option kws autoclose [Boolean] Enable auto-dismiss using timeout.
+  #   Must be true for auto-dismiss to work.
+  # @option kws href     [String] Converts alert to a clickable link.
+  # @option kws action   [String] Stimulus action to fire on click.
+  # @option kws closable [Boolean] Show close button. Set to true to enable
+  #   manual dismissal.
   #
   def initialize(*args, **kws, &block)
     super
 
     @icon = config_option(:icon)
+    @timeout = config_option(:timeout)
+    @autoclose = config_option(:autoclose)
+    @action = config_option(:action)
+    @closable = config_option(:closable, false)
   end
 
   def default_icon_size
@@ -142,7 +185,63 @@ class Daisy::Feedback::AlertComponent < LocoMotion::BaseComponent
   end
 
   def before_render
+    setup_component
+
+    super
+  end
+
+  private
+
+  def setup_component
     add_css(:component, "alert")
     add_html(:component, { role: "alert" })
+
+    setup_stimulus_controller
+    setup_timeout
+    setup_action
+    setup_close_button
+    setup_closable_padding
+  end
+
+  def setup_stimulus_controller
+    return unless closable? || @autoclose
+
+    add_stimulus_controller(:component, "loco-alert")
+  end
+
+  def setup_timeout
+    timeout_value = if @timeout == false
+      nil
+    elsif @timeout.nil?
+      LocoMotion.configuration.default_alert_timeout
+    else
+      @timeout
+    end
+
+    if timeout_value && @autoclose
+      add_html(:component, {
+        "data-loco-alert-timeout-value": timeout_value
+      })
+    end
+  end
+
+  def setup_action
+    if @action
+      add_html(:component, { "data-action": @action })
+    end
+  end
+
+  def setup_close_button
+    return unless closable?
+
+    set_tag_name(:close, :button)
+    add_css(:close, "btn btn-ghost btn-circle btn-xs where:absolute where:top-3 where:right-2")
+    add_html(:close, { "data-action": "click->loco-alert#close" })
+  end
+
+  def setup_closable_padding
+    return unless closable?
+
+    add_css(:component, "where:relative where:pr-10")
   end
 end